add new transport interface
